+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## vNext
 
-- TODO
+- Extensibility updates
+  - Focussing core simulator on OpenAI (moved doc intelligence generator to example extension)
+  - API authorization is now part of forwarders/generators. If you have custom forwarders/generators they need to be updated to handle this (see examples for implementation details)
+
 
 ## v0.3 2024-05-03